Refactors minMaxWorkers from ModelInfo to WorkerPool
This removes some of the data from the ModelInfo to the wlm. Specifically, it
moves tracking of the min and max number of workers.

It also starts turning the wlm.WorkerPool into a public representation of the
management of a single model while the main wlm has all models. As part of this,
some functionality is moved from the main wlm to the pool. In general, we should
eventually not have part of the API be wlm.doSomething(model, ...) and some of
it be wlm.model(model).doSomething(...). I favor the latter approach.

Lastly, this also removes the concept of updating the model and then triggering
model changed. I noticed the change only really affects the scaleModel, so it is
simpler to just move those effects under scaleModel directly.