Multi-device model support
This modifies server to support models with workers on multiple devices rather
than using the previous system of having it as multiple models with versions
indicating the device.

As part of this, it also tries to replace ad-hoc device usages with actual
device. This includes systematically converting device names into Device and
storing device rather than device name or gpu boolean.