Create control plane for workflows (#75)
This updates the management request handler to add a control plane for
workflows. It shares a number of routes between the old models and workflows,
while some routes have different behaviors for each.

Specifically, there is a separate register and list command for both models and
workflows, while all other commands including describe, unregister, and scale
only work on workflows but will still accept "/models/..." routes. The precise
behavior is documented in management_api.md.

There are also a few refactors for ModelServerTest including cleaning up test
client connections, a shared list reader, a shared model registration checking function.

Finally, it also fixes a bug with the WorkLoadManager close not closing the
threadpool. This can result in the tests hanging.