Multi-device model support (#77)
This modifies server to support models with workers on multiple devices rather
than using the previous system of having it as multiple models with versions
indicating the device. The models then share a single work queue. I tested it
with a queue per each device and found this way had slightly better performance.

As part of this, it also tries to replace ad-hoc device usages with actual
device. This includes systematically converting device names into Device and
storing device rather than device name or gpu boolean.

It also makes some changes to model loading. This moves the model loading to a
part of the workflow scaling. Because loading is device dependent, and you can
scale a model on multiple devices, the loading should occur during the scaling
for each device.

The remaining part of installing dependencies, as it is part of serving and not
wlm, has been kept in model registration.