Parameterize WLM (#142)
Right now, the WorkLoadManager operates only in terms of Input and Output.
However, if it is intended to be more general use, then it should be able to use
model and predictors for arbitrary input and output types. This parameterizes
the WLM to make it possible. Within serving, all parameterized types still use
Input and Output and shouldn't behave any differently.