Publish Serving and publish.gradle (#248)
This has two changes. First, it adds the serving to the list of modules to
publish to maven central. The reason for this is that various types of projects
depend on code within serving: plugins, workflow operations, and web clients.

I tried to move the requested components to a new module, but found that almost
the entirety of the serving module was used in one of those three and would have
to be moved. It may be possible to split more out with greater usage of
interfaces, but I think it makes more sense to just distribute serving as a
whole.

The second change made was to create a shared publish.gradle helper similarly to
the one in the main DJL repo. As there were 3 (now four) different modules
published, I figured it was time for adding the full helper.