Create an ensemble workflow (#282)
* Create an ensemble workflow

This creates a test for an ensemble workflow. As part of it, it also modifies
the workflow expression evaluation to return an Item instead of an Input. This
is primarily so that a list can be returned. It is useful for more logical or
control flow functions like map, filter, functionsApply, etc.

It also creates two new built-ins: ensemble and functionsApply.

Lastly, the dependency on Ensembleable means that the current version of DJL
Serving must depend on DJL 0.20.0 (including as snapshot).

* Fix CI

* Fix CI second try