S3 Cache Engine
This creates the S3 Cache engine. It is put into the same cache plugin by
expanding the DDB plugin to handle it as well.

Alongside this, there is some work done to synchronize the efforts on cache
engines. A new BaseCacheEngine class is created to contain some common logic
between the cache engines. The former tests for the DDB cache engine were
generalized a bit and turned into a suite that can be and is run for all of the
three supported cache engines. This ensures (and fixes) some inconsistencies in
behavior among the cache engines.

Also important is that it adds a new test dependency on
localstack (https://localstack.cloud/). This runs a local AWS clone inside a
docker container and is used to verify the running of the S3 cache.