Creates WorkerPoolConfig abstraction
The basic idea is that I replaced the `ModelInfo` inside of the WLM with an
abstraction called `WorkerPoolConfig`. The wpc corresponds to a WorkerPool and
determines what data is saved and work is done within a `WorkerThread`. The main
wpc is still the ModelInfo which extends it.

I also added (and later removed as unused) a LambdaWorkerType that can be used
for other processing. Unlike a WorkflowFunction, it runs in a dedicated worker
thread rather than on the common CompleteableFuture pool.

I think this also helps clean up the abstractions a bit. Within the ModelInfo,
it essentially removed all of the pieces which are more about interacting with
DJL Serving and the WLM more than deep learning. It should also help define
clearer boundaries for when we add expansions and other pieces to the model such
as adapters. I'll also update the architecture section after this (which
probably needed an update regardless).

As future work, it may be cleaner to make the WorkerPoolConfig a final class.
Then, it would include the ModelInfo inside of it as a property rather than
having the ModelInfo extend it. The change was not made now because it adds
additional difficulties when parsing from workerflow.json files.

This also includes updates to the architecture.md that updates the architecture
diagrams and explanation to the current status.