Simplify handling of min/max workers (#1021)
* Simplify handling of min/max workers

This unifies the handling of min/max workers into the two pieces of WorkerGroup
and ModelInfo. The source becomes clear in modelInfo unless a default is passed
during a scaling init of a new worker group. Before, there were more layers of
passed in data and defaults in ways that were difficult to track.

* Checks for 0 min/max workers