Adds adapter support
This PR serves to add adapter support. It is supported with two different
customer experiences which can be seen in the ModelServerTest file. The first is
through the new adapter management API and adding an adapter query parameter to
prediction calls. The second is by creating a workflow with the
AdapterWorkflowFunction.

Both are testing using the adaptecho model. It is a simple python echo model
that prepends the adapter contents. It can be used as a simple example of how
the adapter should function for a model handler. Note the usage of additional
functions register_adapter and unregister_adapter alongside the main handler.

Then, there is a new Adapter abstract class. It only has a PyAdapter
implementation for now. In the future, this should depend on the Patch from DJL
and will require an addition Engine.loadAdapter(name, url) function to create
the engine specific patches. This is left as a future improvement. For now, it
has functionality to both register and unregister the adapter through the
predictor. It also has helpers to integrate the registration and unregistration
flows alongside both the API and workflow approach.

After that, those Adapters are stored in the ModelInfo. It adds a new concept of
configJobs that are run on worker threads. The ModelInfo will create config jobs
to register all adapters on thread creation and to register new adapters on
existing workers.

To support the API approach, the main changes are through the new
AdapterManagementRequestHandler. They are all testing in the ModelServerTest.

To support the workflow approach, it begins by making both workflows and
workflow functions contain both a prepare and close function. Preparing the
AdapterWorkflowFunction will register it and closing it will unregister it. The
test features two workflows with a shared model (adaptecho) that are loaded
together.

Another miscellaneous improvement to mention is the change to Job. Job has two
new arguments. The first is a Job runner that determines what will run. It can
either take the default behavior of executing the predictor through the
ThreadConfig or contain custom behavior (register and unregister). The second
change is that Output is stored in the job rather than returned directly. This
helps deal with some job runners that may be successful for part of the batch
but fail with other parts. It also cleans up the logic with respect to grouping
bu job runner. The JobFunction is also a new functioanal interface to handle
exceptions.