Fixes API responses
This is an update to fix the changes to the workflow response from
https://github.com/deepjavalibrary/djl-serving/pull/1022. Before, we had no test
that checks the response components decided by the names of Java response object
properties (workflows, models, etc.). This adds some new JSON testing to help
ensure that these are not accidentally changed, fixing this issue and hopefully
avoiding future breakages.