Cleans tensorParallelDegree with MultiDevice
This is a refactor to simplify the handling of tensor parallel degree. Before,
it is read independently in 3+ locations in code and the behavior determining
the tpDegree is hard to follow. This moves the reading to a single place and
then represents it using a MultiDevice. This also means that the behavior can be
entirely visible - a worker group will show up with the tp devices that it's
worker will be using.