Fixes changes to hf parse_input signature
This removes the changes to the parse_input signature for adapters for users who
override it.

The fix is a bit of a hack using the obj to pass it instead of the return value.
However, moving the adapter parsing logic outside of the main parse_input would
require redoing much of the decoding work and would therefore increase the
parsing costs.

@frankfliu