[CI IB] Support variables
This adds support for a new section on variables in the IB configurations. They
are specified by

```
[test_name]
test
[vars]
CONCURRENCY={1,16,32}
RANGE=[1..5]
```

This will then produce a template result for each combination of vars (note
ranges are following python semantics of inclusive start and exclusive end).

It can be used to support larger configurations with variations.

I also added a small update to the record benchmarks script to also pre-compute
throughput.