Works with manual initialization
This updates DJL Serving to follow the engine initialization standards set in
https://github.com/deepjavalibrary/djl/pull/2885. First, it updates the
PyEngineProvider to follow the EngineProvider conventions.

It also updates the DependencyManager, fixing some bugs that inspired
https://github.com/deepjavalibrary/djl/pull/2934. The dependency manager before
this change would always re-register the engine with a new provider. Beforehand,
the engine was fully static and unable to re-initialize, so this would not
accidentally re-initialize engines. After this, it would re-initialize engines
that did not support it causing errors. Instead, it should only register new
engines rather than all providers to avoid these accidental re-initialization.