Workflow Templates
This creates a new concept of workflow templates. They are largely the same as
existing workflows, but expand the `WorkflowDefinition` to support template
replacement fields. The workflow templates can be used to simplify the creation
of new workflows, especially ones that follow simple patterns.

As future work, it may be worth considering whether to support:
- Ways to define the templates as a file rather than just through the management
API
- Adding templates using a pre-defined folder or the config