Adds adapter registration options
This adds support for additional options that can be used for adapter
registration. These options will be passed through to the model and can
therefore be model specific. The options are formatted as a Map<String, String>
and can be passed through both HTTP and AdapterWorkflowFunction registration.
Using the adapter directory support does not support options.