[CI] Supports temp repo in execute job
This revives the use of the temp repo from #1266 and implements it in the
execute job. The job can now be used to automatically test a particular branch
through the temp repo. The repo is automatically built as part of the job. So,
to test a branch using the pytest suite with actions, these are the only steps
required:

1. Push the branch to test to the deepjavalibrary/djl-serving upstream repo
2. Run the integration execute job with the default djl-version "temp"